+++ compiler/src/test/fixtures/allscalars/output/java-server/com/example/allscalars/MyRecursiveInput.java
@@ -43,7 +43,7 @@ public final class MyRecursiveInput {
   }
 
   @Override
-  public final boolean equals(final Object o) {
+  public boolean equals(final Object o) {
     if ( this == o ) {
       return true;
     } else if ( !( o instanceof MyRecursiveInput ) ) {
@@ -55,12 +55,12 @@ public final class MyRecursiveInput {
   }
 
   @Override
-  public final int hashCode() {
+  public int hashCode() {
     return Objects.hash( child, v );
   }
 
   @Override
-  public final String toString() {
+  public String toString() {
     return "MyRecursiveInput[child=" + child + ", v=" + v + "]";
   }
 }