+++ compiler/src/test/fixtures/allscalars/output/java-server/com/example/allscalars/MyRecursiveListInput.java
@@ -46,7 +46,7 @@ public final class MyRecursiveListInput {
   }
 
   @Override
-  public final boolean equals(final Object o) {
+  public boolean equals(final Object o) {
     if ( this == o ) {
       return true;
     } else if ( !( o instanceof MyRecursiveListInput ) ) {
@@ -58,12 +58,12 @@ public final class MyRecursiveListInput {
   }
 
   @Override
-  public final int hashCode() {
+  public int hashCode() {
     return Objects.hash( child, v );
   }
 
   @Override
-  public final String toString() {
+  public String toString() {
     return "MyRecursiveListInput[child=" + child + ", v=" + v + "]";
   }
 }