+++ compiler/src/test/fixtures/basic/output/java-server/com/example/basic/ObservationInput.java
@@ -40,7 +40,7 @@ public final class ObservationInput {
   }
 
   @Override
-  public final boolean equals(final Object o) {
+  public boolean equals(final Object o) {
     if ( this == o ) {
       return true;
     } else if ( !( o instanceof ObservationInput ) ) {
@@ -52,12 +52,12 @@ public final class ObservationInput {
   }
 
   @Override
-  public final int hashCode() {
+  public int hashCode() {
     return Objects.hash( type, value );
   }
 
   @Override
-  public final String toString() {
+  public String toString() {
     return "ObservationInput[type=" + type + ", value=" + value + "]";
   }
 }