+++ compiler/src/test/fixtures/coerceid/output/java-server/com/example/coerceid/MyCoercableInput.java
@@ -55,7 +55,7 @@ public final class MyCoercableInput {
   }
 
   @Override
-  public final boolean equals(final Object o) {
+  public boolean equals(final Object o) {
     if ( this == o ) {
       return true;
     } else if ( !( o instanceof MyCoercableInput ) ) {
@@ -67,12 +67,12 @@ public final class MyCoercableInput {
   }
 
   @Override
-  public final int hashCode() {
+  public int hashCode() {
     return Objects.hash( id, otherId );
   }
 
   @Override
-  public final String toString() {
+  public String toString() {
     return "MyCoercableInput[id=" + id + ", otherId=" + otherId + "]";
   }
 }