+++ compiler/src/test/fixtures/coerceid/output/java-server/com/example/coerceid/MyInput.java
@@ -42,7 +42,7 @@ public final class MyInput {
   }
 
   @Override
-  public final boolean equals(final Object o) {
+  public boolean equals(final Object o) {
     if ( this == o ) {
       return true;
     } else if ( !( o instanceof MyInput ) ) {
@@ -54,12 +54,12 @@ public final class MyInput {
   }
 
   @Override
-  public final int hashCode() {
+  public int hashCode() {
     return Objects.hash( id, otherId );
   }
 
   @Override
-  public final String toString() {
+  public String toString() {
     return "MyInput[id=" + id + ", otherId=" + otherId + "]";
   }
 }