+++ compiler/src/test/fixtures/typemapping/output/java-server/com/example/typemapping/ResourceInput.java
@@ -42,7 +42,7 @@ public final class ResourceInput {
   }
 
   @Override
-  public final boolean equals(final Object o) {
+  public boolean equals(final Object o) {
     if ( this == o ) {
       return true;
     } else if ( !( o instanceof ResourceInput ) ) {
@@ -54,12 +54,12 @@ public final class ResourceInput {
   }
 
   @Override
-  public final int hashCode() {
+  public int hashCode() {
     return Objects.hash( name, type );
   }
 
   @Override
-  public final String toString() {
+  public String toString() {
     return "ResourceInput[name=" + name + ", type=" + type + "]";
   }
 }