+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Remove final qualifier on `toString()`, `hashCode()` and `equals(Object)` methods in generated input classes.
+  The final qualifier is unnecessary as the class is final.
+
 ### [v0.01](https://github.com/realityforge/giggle/tree/v0.01) (2019-06-11)
 [Full Changelog](https://github.com/realityforge/giggle/compare/f99bb23e8e6ddeeeb17b79e337f84fdd539308a0...v0.01)